role based authentication done

+++ routes/api.php
@@ -8,19 +8,41 @@ use App\Http\Controllers\UserInfoController;
 use App\Http\Controllers\CommentsController; 
 use App\Http\Controllers\ShapeController;
 use App\Http\Controllers\AuthController;
+use App\Http\Middleware\AdminMiddleware;
 
 
 Route::get('/user', function (Request $request) {
     return $request->user();
 })->middleware('auth:sanctum');
 
+// Syntax for login User Authentication
+// Route::middleware('auth:sanctum')->group(function () {
+    
+// });
+
+// Syntax for User Based Authentication
+//>middleware(AdminMiddleware::class)
+
+
 // Recipe
+Route::middleware('auth:sanctum')->group(function () {
+
 Route::get('recipes',[RecipeController::class,'index']);
 Route::post('recipes',[RecipeController::class,'store']); 
 Route::get('recipes/{id}',[RecipeController::class,'show']);
 Route::get('recipes/{id}/edit',[RecipeController::class,'edit']);
-Route::put('recipes/{id}/edit',[RecipeController::class,'update']);
-Route::delete('recipes/{id}/delete',[RecipeController::class,'destroy']); 
+Route::put('recipes/{id}/edit',[RecipeController::class,'update'])->middleware(AdminMiddleware::class);
+Route::delete('recipes/{id}/delete',[RecipeController::class,'destroy'])->middleware(AdminMiddleware::class); 
+
+});
+
+// Route::get('recipes',[RecipeController::class,'index']);
+// Route::post('recipes',[RecipeController::class,'store']); 
+// Route::get('recipes/{id}',[RecipeController::class,'show']);
+// Route::get('recipes/{id}/edit',[RecipeController::class,'edit']);
+// Route::put('recipes/{id}/edit',[RecipeController::class,'update']);
+// Route::delete('recipes/{id}/delete',[RecipeController::class,'destroy']); 
+
 
 // Category 
 Route::get('categories',[RecipeCategoryController::class,'index']);
